feat: add empty suggestions for forms opportunity - refactoring
